+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Pending
 
 * Display the text version of the 2FA secret code at the same time as the QR code (#190)
+* Change error message feeder command gives when inbox does not exit
+* Fix non-ASCII filename handling in attachment download (#206)
 
 ### Deploy for 2017-05-20